fastapi/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-multipart
sqlalchemy
mysql-connector-python
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
